Fix warning in src/win32/version.h

+++ src/win32/version.h
@@ -18,7 +18,7 @@ GIT_INLINE(int) git_has_win32_version(int major, int minor, int service_pack)
 	version_test.dwOSVersionInfoSize = sizeof(OSVERSIONINFOEX);
 	version_test.dwMajorVersion = major;
 	version_test.dwMinorVersion = minor;
-	version_test.wServicePackMajor = service_pack;
+	version_test.wServicePackMajor = (WORD)service_pack;
 	version_test.wServicePackMinor = 0;
 
 	version_test_mask = (VER_MAJORVERSION | VER_MINORVERSION | VER_SERVICEPACKMAJOR | VER_SERVICEPACKMINOR);
